zeropykvm/epaper.py
"""E-Paper Display Driver for EPD 2in13 V4.

IMPORTANT WARNINGS:

1. PARTIAL REFRESH LIMITATION
   This screen supports partial refresh, but you CANNOT use partial refresh
   continuously. After several partial refreshes, you MUST perform a full
   refresh to clear the screen.

2. POWER MANAGEMENT (CRITICAL)
   The screen MUST NOT remain powered on for extended periods. When not
   refreshing, always put the screen into sleep mode.

3. REFRESH INTERVAL
   Recommended minimum refresh interval is 180 seconds.

4. SLEEP MODE BEHAVIOR
   After entering sleep mode, the screen will ignore any image data.
   You must re-initialize the display before refreshing again.

Note: This is a stub implementation. On actual Raspberry Pi hardware,
this would use the Waveshare EPD library via ctypes or a Python SPI library.
"""

import logging

logger = logging.getLogger(__name__)


class Display:
    """E-Paper display controller.

    On non-RPi systems, this operates as a no-op display that logs all operations.
    """

    def __init__(self, enabled: bool = True):
        self.enabled = enabled
        self.partial_refresh_count = 0

        if not enabled:
            logger.info("E-Paper display disabled")
            return

        try:
            self._init_hardware()
        except Exception as e:
            logger.warning("Failed to initialize e-Paper display: %s", e)
            logger.info("Falling back to disabled mode")
            self.enabled = False

    def _init_hardware(self) -> None:
        """Initialize the e-paper display hardware.

        This would use the Waveshare EPD library on actual hardware.
        On non-Pi systems, this raises an error.
        """
        # Try to import the hardware libraries
        # On non-Pi systems, these won't be available
        try:
            import spidev  # noqa: F401
            import RPi.GPIO  # noqa: F401
        except ImportError:
            raise RuntimeError("E-Paper hardware libraries not available (not running on RPi)")

        # On actual RPi, initialize the display:
        # 1. DEV_Module_Init()
        # 2. EPD_2in13_V4_Init()
        # 3. EPD_2in13_V4_Clear()
        # 4. Create image buffer
        # 5. Draw initial status screen
        # 6. EPD_2in13_V4_Display_Base(image_buffer)
        logger.info("E-Paper display initialized")

    def update_edid_status(self, success: bool) -> None:
        """Update EDID status line.

        Args:
            success: True for OK, False for ERR.
        """
        if not self.enabled:
            return
        status = "OK" if success else "ERR"
        logger.info("E-Paper: EDID status: %s", status)
        self.partial_refresh_count += 1

    def update_hdmi_status(self, success: bool) -> None:
        """Update HDMI status line.

        Args:
            success: True for OK, False for ERR.
        """
        if not self.enabled:
            return
        status = "OK" if success else "ERR"
        logger.info("E-Paper: HDMI status: %s", status)
        self.partial_refresh_count += 1

    def update_hid_status(self, success: bool) -> None:
        """Update HID status line.

        Args:
            success: True for OK, False for ERR.
        """
        if not self.enabled:
            return
        status = "OK" if success else "ERR"
        logger.info("E-Paper: HID status: %s", status)
        self.partial_refresh_count += 1

    def show_status(self, ip: str, port: int) -> None:
        """Show final status screen with IP and port.

        Args:
            ip: IP address string.
            port: Port number.
        """
        if not self.enabled:
            return
        logger.info("E-Paper: Status: %s:%d", ip, port)

    def sleep(self) -> None:
        """Put display to sleep (MUST be called when not refreshing)."""
        if not self.enabled:
            return
        logger.info("E-Paper: Entering sleep mode")

    def shutdown(self) -> None:
        """Hardware shutdown - clear display and release hardware."""
        if not self.enabled:
            return
        logger.info("E-Paper: Shutting down")

    def deinit(self) -> None:
        """Clean up resources."""
        if not self.enabled:
            return
        self.shutdown()
